adding part 4
added a bar chart, a scatter plot and a pie chart to part 4. The bar chart was added to demonstrate the average sleep by stress levels, the scatter plot was added to demonstrate the relationship between daily steps ad quality of sleep and a pie chart was created to demonstrate the sleep disorders present.